requirements.txt
torch==2.10.0+cpu
torchvision==0.25.0+cpu
nibabel>=5.0.0
scipy>=1.10.0
numpy>=1.24.0
pillow>=10.0.0
monai>=1.3.0
hydra-core>=1.3.0
omegaconf>=2.3.0
